Implement covering and darkening color of covered card.
When one card is generated, a random location is assigned to it.
The constructor will check whether the new card overlaps any old cards,
if so, mark the new card covers the old card and the old card be covered
by the new card.
Color darkening is implemented with setStyleSheet.